Add a reference to the document on expansion of labels
Add a reference to a document describing the semantics of the expansion
of labels and the reason why we have to distinguish between Strings
and labels. This is relevant for external repositories and will become
more important once renaming of external repositories will be used.

That proposal itself was approved already in March 2018. Add it for
completeness.