Add a proposal to allow repository rules to handle download failures
For actions in repository rules (initiated by ctx.execute) the exit
code of the action is passed on to the repository rule so that it
can use it to handle the failure of the action or use actions to
probe parts of the environment. For downloads, however, any failure
is fatal and the whole repository rule is aborted. This proposal
argues that allowing rules to handle download failures is a useful
extension of the build API.