Add a design on additional conditions for repository cache lookup
Bazel has a "repository cache", which contains files indexed by
their hash. While a hash is a good way to identify the contents of
a file, there are situations, where the rule author does not think
of the hash as identifying the file and wants to force a redownload
under certain conditions, in case the hash is "wrong". This document
discusses possible implementations.