Mark authentication design as implemented
It was approved in the Starlark meeting at June 6, 2019
and the implementation completed in 26e3396cb341e173b4be5c780cafac3ec386df5.

Also add a clarifying statement that http_archive will by
default honor a `.netrc` file even though this will not
be part of the build API, but instead an implementation
detail of the corresponding Starlark rule.