Add "Standard exec configurations" to review. (#339)
John Cater already reviewed this, and I've started implementing. But we didn't
share this publicly. Doing so now to update the community and solicit wider
feedback.

Note this doc is both generic in its ideas but also focused on Google's repo as
an initial test of "cleaning up a repo". Other orgs would need to consider that
based on their own patterns. Some of the cleanup can't generalize to other orgs
because of that.

Co-authored-by: lberki <lberki@users.noreply.github.com>